refactor: consolidate service wiring and reorganize tests by area
MediaService is now construction and shutdown order and nothing else: the
two topics the daemon answers about itself, ovos.common_play.ping and
opm.audio.query, moved to the bus edge that already owns every other
subscription, and the one-line config-handler call was inlined.

The test tree is now named by area. The batch-named files dissolved into
the area file matching each subject, and NowPlaying, play routing and the
player's MPRIS paths got area files of their own. One shared fixture
replaces the four near-identical mocked-player factories.
